+++ lib/generators/component/component_generator.rb
@@ -9,6 +9,17 @@ class ComponentGenerator < Rails::Generators::NamedBase
     @family_cst = @family.camelize.pluralize # Force plural
     @comp_cst = @comp.camelize # Tolerate singular and plural
 
-    template 'component.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    case @comp_cst
+    when 'Destroy'
+      template 'destroy.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    when 'Edit'
+      template 'edit.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    when 'Form'
+      template 'form.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    when 'New'
+      template 'new.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    else
+      template 'component.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    end
   end
 end